Move `sync::new_sync::sync_bakes` into `sync` directly
